refactor(table): one file per LSM operation
`table/checkpoint.rs` held the stats types, the fault taxonomy, and the
constants, while the checkpoint loop itself lived in `table.rs` — the only
substantial body in a file where every other `Table` method is a delegate.

Split along the operation boundary the rest of the crate already uses:
`table/lsm_stats.rs` owns the `get_lsm_stats` shape and its tests,
`table/checkpoint.rs` owns the loop plus the fault classification it acts
on, and `Table::checkpoint_lsm` becomes a one-line delegate like its
neighbors. The loop's helper constants drop to private now that nothing
outside the module reads them.

No behavior change.
